Extract Kotlin DSL caches soak testing fixtures
for clarity
