+++ subprojects/soak/src/integTest/kotlin/org/gradle/kotlin/dsl/caching/BuildCacheIntegrationTest.kt
@@ -16,6 +16,10 @@
 
 package org.gradle.kotlin.dsl.caching
 
+import org.gradle.kotlin.dsl.caching.fixtures.CachedScript
+import org.gradle.kotlin.dsl.caching.fixtures.cachedBuildFile
+import org.gradle.kotlin.dsl.caching.fixtures.cachedSettingsFile
+import org.gradle.kotlin.dsl.caching.fixtures.compilationCache
 import org.gradle.kotlin.dsl.fixtures.normalisedPath
 
 import org.gradle.soak.categories.SoakTest
